Exposes `ParsedType` as public api.
Motivation is described in #1567.

Closes #1567
