+++ docs/reference/index.rst
@@ -39,3 +39,4 @@ API reference
     template
     testing
     types
+    typing
