+++ docs/reference/typing.rst
@@ -1,0 +1,7 @@
+typing
+======
+
+.. py:currentmodule:: litestar.typing
+
+.. automodule:: litestar.typing
+   :members:
